+++ src/collapse/collapse-context.ts
@@ -6,6 +6,7 @@ interface CollapseContextInterface {
   collapsed: boolean;
   duration: number;
   disableAnimation: boolean;
+  keepMounted?: boolean;
   setCollapsed: () => void;
   id: string;
 }
@@ -14,6 +15,7 @@ export const CollapseContext = createContext<CollapseContextInterface>({
   collapsed: true,
   duration: BASE_ANIMATION_DURATION,
   disableAnimation: false,
+  keepMounted: false,
   setCollapsed: () => {},
   id: '',
 });